+++ CHANGELOG.md
@@ -25,6 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## `bitrouter` - [0.2.2](https://github.com/bitrouter/bitrouter/compare/v0.2.0...v0.2.2) - 2026-03-10
 
 ### Added
+- *(runtime)* add basic api key auth ([#39](https://github.com/bitrouter/bitrouter/pull/39))
+
+## `bitrouter` - [0.2.2](https://github.com/bitrouter/bitrouter/compare/v0.2.0...v0.2.2) - 2026-03-10
+
+### Added
 - *(runtime)* basic auth
 
 ### Other